Refactor Data Quality Tests and Update TableClass Data Types
- Updated Data Quality test cases to utilize a new utility function for visiting the create test case panel from the entity page, reducing code duplication.
- Modified the data type of the second column in TableClass from 'NUMERIC' to 'INT' for consistency.
- Enhanced data quality utility functions to include creating and editing test cases, improving test case management.

+++ openmetadata-ui/src/main/resources/ui/playwright/support/entity/TableClass.ts
@@ -155,8 +155,8 @@ export class TableClass extends EntityClass {
       },
       {
         name: this.columnsName[1],
-        dataType: 'NUMERIC',
-        dataTypeDisplay: 'numeric',
+        dataType: 'INT',
+        dataTypeDisplay: 'int',
         description:
           'The ID of the store. This column is a foreign key reference to the shop_id column in the dim.shop table.',
       },
